requirements.txt
Flask==3.1.1
qrcode==8.2
pillow==11.2.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
cryptography==45.0.4
pytest==8.4.1